+++ packages/app-webdir-ui/src/WebDirectoryComponent/index.js
@@ -5,6 +5,7 @@ import React, { useState, useEffect } from "react";
 
 import trackReactComponent from "../../../../shared/services/componentDatalayer";
 import FacultyRankTabPanels from "../FacultyRankComponent";
+import { FILTER_CHOICES } from "../helpers/constants";
 import { FilterComponent } from "../helpers/Filter";
 import { engineNames, engines } from "../helpers/search";
 import { SortPicker } from "../SearchPage/components/sort";
@@ -15,7 +16,7 @@ import { WebDirLayout, FacultyRankLayout } from "./index.styles";
  * React component for displaying web directory search results.
  *
  * @param {Object} props - The props for configuring the WebDirectory component.
- * @param {string} props.searchType - The type of web directory search (e.g., 'departments', 'people').
+ * @param {"departments" | "faculty_rank" | "people" | "people_departments"} props.searchType - The type of web directory search (e.g., 'departments', 'people').
  * @param {string} props.ids - The IDs used for searching (e.g., department IDs, ASURITE IDs).
  * @param {string} props.deptIds - The department IDs for searching.
  * @param {string} props.API_URL - The API URL for performing the search.
@@ -162,34 +163,7 @@ function WebDirectory({
           {alphaFilter === "true" && (
             <FilterComponent
               filterLabel="Filter By Last Initial"
-              choices={[
-                "A",
-                "B",
-                "C",
-                "D",
-                "E",
-                "F",
-                "G",
-                "H",
-                "I",
-                "J",
-                "K",
-                "L",
-                "M",
-                "N",
-                "O",
-                "P",
-                "Q",
-                "R",
-                "S",
-                "T",
-                "U",
-                "V",
-                "W",
-                "X",
-                "Y",
-                "Z",
-              ]}
+              choices={FILTER_CHOICES}
               onChoose={filterLetter =>
                 setRequestFilters({ ...requestFilters, lastInit: filterLetter })
               }
@@ -229,6 +203,7 @@ function WebDirectory({
       <FacultyRankTabPanels
         {...enginesWithParams[searchTypeEngineMap[searchType]]}
         alphaFilter={alphaFilter}
+        filters={requestFilters}
       />
     </FacultyRankLayout>
   );