+++ packages/component-header-footer/.storybook/preview.js
@@ -1,0 +1,15 @@
+const parameters = {
+  actions: { argTypesRegex: "^on[A-Z].*" },
+  percy: {
+    include: [
+      "ASU Header: Default\$",
+    ],
+  },
+};
+
+/** @type { import('@storybook/react').Preview } */
+const preview = {
+  parameters
+};
+
+export default preview;