+++ packages/component-header/src/components/HeaderMain/Partner/index.js
@@ -19,7 +19,17 @@ const Partner = () => {
     <PartnerLogosWrapper breakpoint={breakpoint} data-testid="partner">
       <a
         href={partnerLogo?.brandLink ?? "https://starbucks.asu.edu/"}
-        onFocus={() => trackGAEvent({ text: "starbucks logo" })}
+        onClick={() =>
+          trackGAEvent({
+            event: "link",
+            action: "click",
+            name: "onclick",
+            region: "navbar",
+            type: "internal link",
+            section: "main navbar",
+            text: "partner logo",
+          })
+        }
       >
         <img
           src={partnerLogo?.src ?? starbucksLogo}
@@ -34,7 +44,7 @@ const Partner = () => {
       </a>
       <a
         href={asuLogo?.brandLink ?? "https://asu.edu"}
-        onFocus={() => trackGAEvent({ text: "asu logo" })}
+        onClick={() => trackGAEvent({ text: "asu logo" })}
       >
         <img
           className="vert"