+++ packages/components-core/src/components/RankingCard/index.stories.js
@@ -3,6 +3,8 @@ import React from "react";
 
 import { RankingCard } from ".";
 
+import { imageAny } from "../../../../../shared/assets";
+
 export default {
   title: "UDS/Ranking Card",
   component: RankingCard,
@@ -21,7 +23,7 @@ const Template = args => (
 export const Large = Template.bind({});
 Large.args = {
   imageSize: "large",
-  image: "https://source.unsplash.com/WLUHO9A_xik/600x400",
+  image: imageAny(),
   imageAlt: "Image alt text",
   heading: "Ranking title goes here, under the photo",
   body: "ASU has topped U.S. News & World Report’s “Most Innovative Schools list since the inception of the category in 2016. ASU again placed ahead of Stanford and MIT on the list, based on a survey of peers. College presidents, provosts and admissions deans around the country nominated up to 10 colleges or universities that are making the most innovative improvements.",
@@ -31,7 +33,7 @@ Large.args = {
 export const Small = Template.bind({});
 Small.args = {
   imageSize: "small",
-  image: "https://source.unsplash.com/WLUHO9A_xik/600x400",
+  image: imageAny(),
   imageAlt: "Image alt text",
   heading: "Ranking title goes here, under the photo",
   body: "ASU has topped U.S. News & World Report’s “Most Innovative Schools list since the inception of the category in 2016. ASU again placed ahead of Stanford and MIT on the list, based on a survey of peers. College presidents, provosts and admissions deans around the country nominated up to 10 colleges or universities that are making the most innovative improvements.",