+++ packages/unity-react-core/src/components/ComponentCarousel/core/utils/init-carousel.js
@@ -1,7 +1,7 @@
 /* eslint-disable quote-props */
 // @ts-check
 import React from "react";
-import ReactDOM from "react-dom";
+import { createRoot } from "react-dom/client";
 
 import {
   CardCarousel,
@@ -11,7 +11,8 @@ import {
 } from "../../components";
 
 const RenderPreact = (component, props, target) => {
-  ReactDOM.render(React.createElement(component, props), target);
+  const root = createRoot(target);
+  root.render(React.createElement(component, props));
 };
 
 /**